fix user-review-service bugs

+++ user-review-service/src/main/java/com/n11/userreviewservice/repository/ReviewRepository.java
@@ -5,12 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.util.List;
 
 public interface ReviewRepository extends JpaRepository<Review,Long> {
 
     @Query("SELECT SUM(r.score) FROM Review r WHERE r.restaurantId = :restaurantId")
     Double findAllRateByRestaurantId(@Param("restaurantId") String restaurantId);
 
-    Long countByRestaurantId(String s);
+    int countByRestaurantId(String s);
 }
